Add files via upload
This is the preliminary progress page. Essentially the button triggers the message on the screen to change. To integrate this into the larger system I recommend basically commenting out line 70 to hide the button and finding a way to trigger it from an outside class as the initial message shows Ready to Cook, after one press it shows Cooking, and after a second press it shows Ready. additionally I was able to get the logo to load on my computer but ideally we need to find a way to get it to work wherever as my use required full path to my home directory.